requirements.txt
flask
flask-cors
pypdf2
langchain
openai
faiss-cpu
python-dotenv
tiktoken
langchain-community
